moved style files for index.js and App.js to styles folder
