+++ client/src/styles/App.scss
@@ -1,5 +1,5 @@
-@import './styles/fonts.module.scss';
-@import './styles/colors.module.scss';
+@import './fonts.module.scss';
+@import './colors.module.scss';
 
 * {
   margin: 0;
