+++ client/src/styles/index.scss
@@ -1,5 +1,5 @@
-@import './styles/colors.module.scss';
-@import './styles/fonts.module.scss';
+@import './colors.module.scss';
+@import './fonts.module.scss';
 
 * {
   margin: 0;
